requirements.txt
bounded-pool-executor
requests
confluent-kafka
avro-python3
redis
loguru
mysql-connector-python
lxml
bs4